+++ test/dummy/app/assets/config/manifest.js
@@ -2,5 +2,6 @@
 //= link_directory ../javascripts .js
 //= link_directory ../stylesheets .css
 //= link flat_pack/variables.css
+//= link flat_pack/application.css
 //= link flat_pack/rich_text.css
 //= link tailwind.css